The number of reported ESLint errors goes down to 579
This simple modification of legacy .eslintrc config files in the folders
with old code drops the number of reported ESLint erros down to 579. It
seems that remaining errors should be just fixed, as they come from code
which does not comply nor with the AirBnB ReactJS style, which should be
used for all new code, nor with the old code rules.